Formatted Code, Renamed Folder
Formatted the test code, renamed the folder holding the tests for tripoverviewpage to match the viewmodel name. Removed unused error messages.
