+++ code/CapstoneTest/DesktopTests/ViewModels/TestTripOverviewPage/TestConstructor.cs
@@ -5,7 +5,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using ReactiveUI;
 
-namespace CapstoneTest.DesktopTests.ViewModels.TestTripOverviewWindow
+namespace CapstoneTest.DesktopTests.ViewModels.TestTripOverviewPage
 {
     [TestClass]
     public class TestConstructor
