+++ code/CapstoneTest/DesktopTests/ViewModels/TestTripOverviewPage/TestUpdateWaypoints.cs
@@ -6,7 +6,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using ReactiveUI;
 
-namespace CapstoneTest.DesktopTests.ViewModels.TestTripOverviewWindow
+namespace CapstoneTest.DesktopTests.ViewModels.TestTripOverviewPage
 {
     [TestClass]
     public class TestUpdateWaypoints
